rename classic to classic_api
