+++ search/factory.py
@@ -8,7 +8,7 @@ from flask_s3 import FlaskS3
 from arxiv.base import Base
 from arxiv.base.middleware import wrap, request_logs
 from arxiv.users import auth
-from search.routes import ui, api, classic
+from search.routes import ui, api, classic_api
 from search.services import index
 from search.converters import ArchiveConverter
 from search.encode import ISO8601JSONEncoder
@@ -85,12 +85,12 @@ def create_classic_api_web_app() -> Flask:
 
     Base(app)
     auth.Auth(app)
-    app.register_blueprint(classic.blueprint)
+    app.register_blueprint(classic_api.blueprint)
 
     wrap(app, [request_logs.ClassicLogsMiddleware,
                auth.middleware.AuthMiddleware])
 
-    for error, handler in classic.exceptions.get_handlers():
+    for error, handler in classic_api.exceptions.get_handlers():
         app.errorhandler(error)(handler)
 
     return app
