+++ search/routes/classic_api/exceptions.py
