+++ search/routes/classic_api/tests/test_classic.py
@@ -34,7 +34,7 @@ class TestClassicAPISearchRequests(TestCase):
 
     def test_request_without_token(self):
         """No auth token is provided on the request."""
-        response = self.client.get('/classic/query?search_query=au:copernicus')
+        response = self.client.get('/classic_api/query?search_query=au:copernicus')
         self.assertEqual(response.status_code, HTTPStatus.UNAUTHORIZED)
 
     def test_with_token_lacking_scope(self):
@@ -42,11 +42,11 @@ class TestClassicAPISearchRequests(TestCase):
         token = helpers.generate_token('1234', 'foo@bar.com', 'foouser',
                                        scope=[Scope('something', 'read')])
         response = self.client.get(
-            '/classic/query?search_query=au:copernicus',
+            '/classic_api/query?search_query=au:copernicus',
             headers={'Authorization': token})
         self.assertEqual(response.status_code, HTTPStatus.FORBIDDEN)
 
-    @mock.patch(f'{factory.__name__}.classic.classic')
+    @mock.patch(f'{factory.__name__}.classic_api.classic_api')
     def test_with_valid_token(self, mock_controller):
         """Client auth token has required public read scope."""
         document = dict(
@@ -106,11 +106,11 @@ class TestClassicAPISearchRequests(TestCase):
         r_data = {'results': docs, 'query': domain.ClassicAPIQuery(id_list=['1234.5678'])}
         mock_controller.query.return_value = r_data, HTTPStatus.OK, {}
         response = self.client.get(
-            '/classic/query?search_query=au:copernicus',
+            '/classic_api/query?search_query=au:copernicus',
             headers=self.auth_header)
         self.assertEqual(response.status_code, HTTPStatus.OK)
 
-    @mock.patch(f'{factory.__name__}.classic.classic')
+    @mock.patch(f'{factory.__name__}.classic_api.classic_api')
     def test_paper_retrieval(self, mock_controller):
         """Test single-paper retrieval."""
         document = dict(
@@ -170,7 +170,7 @@ class TestClassicAPISearchRequests(TestCase):
         r_data = {'results': docs, 'query': domain.APIQuery()}
         mock_controller.paper.return_value = r_data, HTTPStatus.OK, {}
         response = self.client.get(
-            '/classic/1234.56789v6', headers=self.auth_header
+            '/classic_api/1234.56789v6', headers=self.auth_header
         )
         self.assertEqual(response.status_code, HTTPStatus.OK)
 
@@ -199,7 +199,7 @@ class TestClassicAPISearchRequests(TestCase):
 
     def test_start_not_a_number(self):
         response = self.client.get(
-            '/classic/query?search_query=au:copernicus&start=non_number',
+            '/classic_api/query?search_query=au:copernicus&start=non_number',
             headers=self.auth_header)
         self.assertEqual(response.status_code, HTTPStatus.BAD_REQUEST)
         self.check_validation_error(
@@ -210,7 +210,7 @@ class TestClassicAPISearchRequests(TestCase):
 
     def test_start_negative(self):
         response = self.client.get(
-            '/classic/query?search_query=au:copernicus&start=-1',
+            '/classic_api/query?search_query=au:copernicus&start=-1',
             headers=self.auth_header)
         self.assertEqual(response.status_code, HTTPStatus.BAD_REQUEST)
         self.check_validation_error(
@@ -221,7 +221,7 @@ class TestClassicAPISearchRequests(TestCase):
 
     def test_max_results_not_a_number(self):
         response = self.client.get(
-            '/classic/query?search_query=au:copernicus&max_results=non_number',
+            '/classic_api/query?search_query=au:copernicus&max_results=non_number',
             headers=self.auth_header)
         self.assertEqual(response.status_code, HTTPStatus.BAD_REQUEST)
         self.check_validation_error(
@@ -232,7 +232,7 @@ class TestClassicAPISearchRequests(TestCase):
 
     def test_max_results_negative(self):
         response = self.client.get(
-            '/classic/query?search_query=au:copernicus&max_results=-1',
+            '/classic_api/query?search_query=au:copernicus&max_results=-1',
             headers=self.auth_header)
         self.assertEqual(response.status_code, HTTPStatus.BAD_REQUEST)
         self.check_validation_error(
@@ -243,7 +243,7 @@ class TestClassicAPISearchRequests(TestCase):
 
     def test_invalid_arxiv_id(self):
         response = self.client.get(
-            '/classic/query?id_list=cond—mat/0709123',
+            '/classic_api/query?id_list=cond—mat/0709123',
             headers=self.auth_header)
         self.assertEqual(response.status_code, HTTPStatus.BAD_REQUEST)
         self.check_validation_error(
